Rename adapters to clarify their domains and resolve collision
Renamed the newly extracted materials adapter to `ReferencesAdapter.kt` to align with `ReferencesFragment`. Renamed the conflicting contacts adapter used in `AchievementFragment` to `AchievementsAdapter.kt` and moved it to the `ui/user` package to resolve the naming collision and clarify its distinct purpose. Tests pass.
